% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spillover-package.R
\docType{data}
\name{dy2009}
\alias{dy2009}
\title{Diebold and Yilmaz (2009) dataset}
\format{
a \code{data.frame}-class dataset
}
\description{
A dataset consisting of 829 weekly observations of Global Stock Market Returns, 10/1/1992 -- 23/11/2007.
Find details in Diebold and Yilmaz (2009)
}
\examples{
data(dy2009)
head(dy2009)  
summary(dy2009)  # Same as Diebold and Yilmaz (2012) summary statistics
}
\references{
Diebold, F. X. & Yilmaz, K. (2009). \emph{Measuring Financial Asset Return and Volatility Spillovers, with Application to Global Equity Markets}. The Economic Journal, 119, 158--171
}
\keyword{dataset}
