% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spillover-package.R
\docType{data}
\name{dy2012}
\alias{dy2012}
\title{Diebold and Yilmaz (2012) dataset}
\format{
a \code{data.frame}-class dataset
}
\description{
A dataset consisting of 2771 log volatility daily observations of 4 variables: 
Stocks (SP500), Bonds (R_10Y), Commodities (DJUBSCOM) and  FX (USDX).
The period for this dataset is from Jan 25, 1999 to Jan 29, 2010.
}
\examples{
data(dy2012)
head(dy2012)  
summary(dy2012)  # Same as Diebold and Yilmaz (2012) summary statistics
}
\references{
Diebold, F. X. & Yilmaz, K.(2012). \emph{Better to Give than to Receive: Predictive Directional Measurement of Volatility Spillovers}. International Journal of Forecasting. 28, 57–66.
}
\keyword{dataset}
