% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{net}
\alias{net}
\title{Net spillovers}
\usage{
net(x)
}
\arguments{
\item{x}{Object of class \sQuote{\code{spillover.table}} generated by either \code{O.spillover()} or  \code{G.spillover()}.}
}
\value{
A list length \emph{K} holding the generalized forecast error variances as matrices.
}
\description{
Computes the net spillover index.
}
\references{
Pesaran, M. H. and Shin, Y. (1998). \emph{Generalized impulse response analysis in linear multivariate models}. Economics Letters, 58(1):17-29.
}
\seealso{
\code{\link{O.spillover}} \code{\link{G.spillover}}
}
\author{
Jilber Urbina
}
