% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spillover-package.R
\docType{data}
\name{rol.vol}
\alias{rol.vol}
\title{Two-days Rolling Average Intra-day Volatilities}
\format{
a \code{zoo}-class dataset
}
\description{
A dataset of class \code{zoo} consisting of 1633 two-days rolling average observations on 
intraday volatilities based on Garman and Klass (1980) for six leading stock indices: 
S&P 500 (US), FTSE 100 (UK), EURO STOXX 50 (Eurozone), BOVESPA (Brazil), NIKKEI 225 (Japan) and
S&P ASX 200 (Australia). EURO STOXX 50 covers 50 stocks from 12 Eurozone countries: Austria, Belgium, 
Finland, France, Germany, Greece, Ireland, Italy, Luxembourg, the Netherlands, Portugal and Spain.
The period for this dataset is from June 13, 2003 to September 15, 2009. All series are in US Dollars.
}
\examples{
data(rol.vol)
head(rol.vol)  # First 6 observations
tail(rol.vol)  # Last 6 observations
}
\keyword{dataset}
