\name{Spillover-package}
\alias{Spillover-package}
\alias{Spillover}
\docType{package}
\title{
\packageTitle{Spillover}
}
\description{
\packageDescription{Spillover}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Spillover}
\packageIndices{Spillover}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Spillover}

Maintainer: \packageMaintainer{Spillover}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
