% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spillover-package.R
\docType{package}
\name{Spillover}
\alias{Spillover}
\alias{Spillover-package}
\title{Spillover}
\description{
Spillover index based on both, orthogonalized and generalized forecast error variance decomposition of a VAR(p) for n step ahead forecast. It computes both orthogonalized and generalized directional spillover indices proposed by Diebold and Yilmaz (2009, 2012)
}
\references{
Diebold, F. X. & Yilmaz, K. (2009). \emph{Measuring Financial Asset Return and Volatility Spillovers, with Application to Global Equity Markets}. The Economic Journal, 119, 158-171.
 
 Diebold, F. X. & Yilmaz, K.(2012). \emph{Better to Give than to Receive: Predictive Directional Measurement of Volatility Spillovers}. International Journal of Forecasting. 28, 57–66.
 
 Hamilton, J. (1994), \emph{Time Series Analysis}, Princeton University Press, Princeton.
 
 Lutkepohl, H. (2006), \emph{New Introduction to Multiple Time Series Analysis}, Springer, New York.  
 
 Pesaran, M. H. and Shin, Y. (1998). \emph{Generalized impulse response analysis in linear multivariate models}. Economics Letters, 58(1):17-29.
}
\author{
Jilber Urbina
}
\keyword{package}
