% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spillover-package.R
\docType{data}
\name{stock.prices}
\alias{stock.prices}
\title{Daily Stock Prices}
\format{
a \code{zoo}-class dataset
}
\description{
A dataset consisting of 3507 daily observations on closed price for six leading stock indices: 
S&P 500 (US), FTSE 100 (UK), EURO STOXX 50 (Eurozone), BOVESPA (Brazil), NIKKEI 225 (Japan) and
S&P ASX 200 (Australia). EURO STOXX 50 covers 50 stocks from 12 Eurozone countries: Austria, Belgium, 
Finland, France, Germany, Greece, Ireland, Italy, Luxembourg, the Netherlands, Portugal and Spain.
The period for this dataset is from December 31, 1999 to June 10, 2013. All series are in US Dollars.
}
\examples{
data(stock.prices)
head(stock.prices)  # First 6 observations
tail(stock.prices)  # Last 6 observations
}
\keyword{dataset}
